fix private record jit
