+++ integration_tests/latest_jdk_tests/src/test/java/io/fury/integration_tests/RecordSerializersTest.java
@@ -219,4 +219,16 @@ public class RecordSerializersTest {
       Assert.assertEquals(fury.deserialize(bytes), o2);
     }
   }
+
+  @Test
+  public void testPrivateRecord() {
+    Fury fury = Fury.builder().build();
+    fury.register(PrivateRecord.class);
+    byte[] serialized = fury.serialize(new PrivateRecord("foo")); // fails
+    Object deserialized = fury.deserialize(serialized);
+    System.out.println(deserialized);
+  }
+
+  private record PrivateRecord(String foo) {
+  }
 }
