+++ java/fury-core/src/main/java/io/fury/util/unsafe/_JDKAccess.java
@@ -261,7 +261,11 @@ public class _JDKAccess {
   }
 
   public static Tuple2<Class<?>, String> getterMethodInfo(Class<?> type) {
-    return methodMap.get(type);
+    Tuple2<Class<?>, String> info = methodMap.get(type);
+    if (info == null) {
+      return  Tuple2.of(Function.class, "apply");
+    }
+    return info;
   }
 
   public static Object makeGetterFunction(
